+++ kopy-args/jvm/main/kotlin/com/javiersc/kotlin/kopy/args/KopyArgument.jvm.kt
@@ -1,5 +1,0 @@
-package com.javiersc.kotlin.kopy.args
-
-import java.io.Serializable
-
-public actual interface KopyArgument : Serializable