+++ kopy-compiler/main/kotlin/com/javiersc/kotlin/kopy/compiler/utils/fqName.kt
@@ -1,0 +1,5 @@
+package com.javiersc.kotlin.kopy.compiler.utils
+
+import org.jetbrains.kotlin.name.FqName
+
+internal fun String.toFqName(): FqName = FqName(this)